refactor(frontend): extract SearchResults component, drop models from home search
Extract shared search logic (data loading, filtering, grouped card
display for titles, manufacturers, people) into a reusable SearchResults
component. Both the home page and /search page now use it, removing
~650 lines of duplication. Models, systems, series, and franchises are
no longer shown in search results on either page.
